fix : eslint 에러 수정
- dependency cycle detected 에러 수정
- types.ts 파일을 추가, 여러 곳에서 import 하고있는 타입을 따로 분리해 참조 문제 해결
