+++ frontend/src/utils/atoms.ts
@@ -1,6 +1,6 @@
 import { atom } from 'recoil'
 import { recoilPersist } from 'recoil-persist'
-import { FilterType } from '../components/recruitment/MultipleFilter'
+import { FilterType } from './types'
 
 const { persistAtom } = recoilPersist()
